refactor(search): split SearchService into retrieval + rerank stages (max-params 2/3)
SearchService injected 7 deps (surreal, embedder, reranker, predicateRouter,
crossEncoder, calibration, metrics). The pipeline stages were already
pre-extracted as private methods; group them behind two stage services:
- SearchRetrievalService (embedder, predicateRouter, calibration) — vector
  + lexical legs with fusion, the predicate/type router, and
  calibrated scoring + per-entity bucketing.
- SearchRerankService (reranker, crossEncoder, metrics) — cross-encoder
  windowing, margin-skip, and the LLM reranker with neighbour injection
  (all rerank metrics live here).
SearchService (surreal, retrieval, rerank) keeps the orchestration: db
lifecycle, identity-merge, edge expansion, PPR, backfill, graph retrieval
— the db-threading stages that need no AI service. PipelineContext moved
to its own module. Each class ≤3 deps; public search()/graphRetrieve()/
expandEntityIdsViaEdges() contracts unchanged.

Part of the max-params=3 god-class split program.
